Bug fix to AFSAuth
When called via Launchd, the subprocess call to aklog was exiting with
the error:
     Error: [Errno 2] No such file or directory
which was due the the binary being in `/usr/local/bin` and the path not
being set appropriately.

This commit fixes by adding an option of aklog_path (defaulting to the
above), along with some debugging output when called with a suitable
verbosity (which helped to trace the source of the error).