Update ErlangURLProvider to only pick .dmg files
The overnight build of Erlang is failing, as it is picking index 0
from the list of Mac OS X 10.10 packages which currently happens to be
a .tgz file, and not the expected .dmg file.

This select breaks the code signature checking, so this patch ensures
we pick the first .dmg file when selecting the latest.