Add receipts version and installs list to the pkginfo
The official SML/NJ package for Mac comes without a valid version number,
so Munki does not know whether the package has been updated or not.  To
fix this, we add the correct version number to the "receipts" dict in the
package info (this lets Munki know it should build a new version when the
version number changes) and a version-specific installed file to the
"installs" dict (this lets Munki know whether or not this version of the
package is installed on the target machine).  Each version of SML/NJ comes
with a versioned README for this package, so we can check for the existence
of that file.  This will break if they ever release a version without this
documentation file.